move action definitions to a separate file so they can be used client-side
